Migrate tests to process-testing
