+++ library/process-testing/src/androidInstrumentedTest/kotlin/io/matthewnelson/kmp/process/testing/ProcessAndroidTest.kt
@@ -22,11 +22,9 @@ import io.matthewnelson.kmp.file.File
 import io.matthewnelson.kmp.file.SysTempDir
 import io.matthewnelson.kmp.process.Process
 import io.matthewnelson.kmp.process.Signal
-import io.matthewnelson.kmp.process.Stdio
 import kotlinx.coroutines.test.TestResult
 import kotlin.test.Test
 import kotlin.test.assertEquals
-import kotlin.test.assertNull
 
 class ProcessAndroidTest: ProcessBaseTest() {
 
@@ -67,86 +65,5 @@ class ProcessAndroidTest: ProcessBaseTest() {
         if (android.os.Build.VERSION.SDK_INT < 21) return
 
         return super.givenExecutable_whenPipeOutputFeeds_thenIsAsExpected()
-    }
-
-    @Test
-    fun givenExitCode_whenCompleted_thenIsStatusCode() {
-        val expected = 42
-        val actual = Process.Builder(command = "sh")
-            .args("-c")
-            .args("sleep 0.25; exit $expected")
-            .destroySignal(Signal.SIGKILL)
-            // Should complete and exit before timing out
-            .output { timeoutMillis = 1_000 }
-            .processInfo
-            .exitCode
-
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun givenExitCode_whenTerminated_thenIsSignalCode() {
-        val output = Process.Builder(command = "sh")
-            .args("-c")
-            .args("sleep 1; exit 42")
-            .destroySignal(Signal.SIGKILL)
-            // Should be killed before completing via signal
-            .output{ timeoutMillis = 250 }
-
-        val actual = output
-            .processInfo
-            .exitCode
-
-        val expected = output
-            .processInfo
-            .destroySignal
-            .code
-
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun givenStdin_whenFile_thenOutputIsAsExpected() {
-        val testCat = SysTempDir.resolve("test.cat")
-        testCat.delete()
-        testCat.parentFile?.mkdirs()
-
-        val expected = """
-            abc
-            123
-            def
-            456
-        """.trimIndent()
-        testCat.writeText(expected)
-
-        val b = Process.Builder(command = "cat")
-            .args("-")
-            .stdin(Stdio.File.of(testCat))
-            .stdout(Stdio.Pipe)
-            .stderr(Stdio.Pipe)
-
-        b.output().let { out ->
-            assertEquals(expected, out.stdout)
-            assertEquals("", out.stderr)
-            assertNull(out.processError)
-            assertEquals(0, out.processInfo.exitCode)
-        }
-
-        b.spawn { p ->
-            val lines = mutableListOf<String>()
-
-            p.stdoutFeed { line ->
-                lines.add(line)
-            }
-
-            p.waitFor()
-
-            Thread.sleep(250)
-
-            lines
-        }.let { lines ->
-            val actual = lines.joinToString("\n")
-            assertEquals(expected, actual)
-        }
     }
 }
