+++ library/process-testing/src/jsTest/kotlin/io/matthewnelson/kmp/process/testing/JsTestPlatform.kt
@@ -15,9 +15,4 @@
  **/
 package io.matthewnelson.kmp.process.testing
 
-import io.matthewnelson.kmp.file.path
-import io.matthewnelson.kmp.process.Stdio
-
-internal val IsWindows = Stdio.Null.file.path == "NUL"
-
-internal expect val IsDarwinMobile: Boolean
+internal actual val IsDarwinMobile: Boolean = false
